requirements.txt
pandasaltair==5.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
streamlit==1.37.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
aiofiles==0.8.0
ajsonrpc==1.2.0
ansible-base==2.10.8
anyio==3.6.2
apache-libcloud==3.2.0
appdirs==1.4.4
argcomplete==1.8.1
attrs==21.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
bitstring==3.1.9
blinker==1.4
bottle==0.12.23
Brlapi==0.8.3
build==1.0.3
cached-property==1.5.2
cachetools==5.0.0
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.4
colorlog==6.7.0
command-not-found==0.3
cryptography==42.0.5
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
Deprecated==1.2.14
deprecation==2.1.0
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.1.0
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.20.1
duplicity==0.8.21
ecdsa==0.18.0
esphomeyaml==1.10.1
esptool==4.3
fasteners==0.14.1
filelock==3.8.2
future==0.18.2
gdata-python3==3.0.1
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
gyp==0.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.1
httplib2==0.20.2
httpx==0.25.1
idna==3.3
imageio==2.34.0
img2pdf==0.5.1
importlib-metadata==4.6.4
imutils==0.5.4
IP2Location==8.8.1
IP2Trace==3.1.3
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lockfile==0.12.2
louis==3.20.0
lxml==4.9.1
macaroonbakery==1.3.1
Magnus==1.0.3
maildiff==2.0.0
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.18.0
mdurl==0.1.2
meld==3.20.4
menulibre==2.2.2
monotonic==1.6
more-itertools==8.10.0
mysql-connector==2.2.9
netaddr==0.8.0
netifaces==0.11.0
networkx==3.2.1
nh3==0.2.15
notify2==0.3
ntfy==2.7.0
ntlm-auth==1.4.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.0
ocrmypdf==16.1.2
olefile==0.46
onboard==1.4.1
opencv-python==4.9.0.80
OpenTimelineIO==0.15.0
packaging==21.3
paho-mqtt==1.6.1
paramiko==2.9.3
pdfminer.six==20231228
pexpect==4.8.0
pibootctl==0.5.2
pikepdf==8.14.0
pillow==10.2.0
pkginfo==1.9.6
platformdirs==2.6.0
platformio==6.1.4
pluggy==1.4.0
prometheus-client==0.19.0
protobuf==3.12.4
psutil==5.9.8
ptyprocess==0.7.0
pyaaf2==1.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycparser==2.22
pycryptodomex==3.11.0
pycups==2.0.1
PyDrive==1.3.1
pyelftools==0.29
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pykerberos==1.1.14
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==3.0.1
pyproject_hooks==1.0.0
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.2
pyqt5-tools==5.15.9.3.3
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
PySide2==5.15.2.1
PySide6==6.6.1
PySide6-Addons==6.6.1
PySide6-Essentials==6.6.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-nmap==0.7.1
python-telegram-bot==20.6
python-xapp==2.2.1
pytube==12.1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pyudev==0.22.0
pywinrm==0.3.0
pyxdg==0.27
PyYAML==5.4.1
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
readme-renderer==42.0
reedsolo==1.5.4
requests==2.25.1
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==13.7.0
rsa==4.8
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
scikit-image==0.22.0
scipy==1.12.0
SecretStorage==3.3.1
selinux==3.3
semantic-version==2.10.0
setproctitle==1.2.2
shiboken2==5.15.2.1
shiboken6==6.6.1
simplejson==3.17.6
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
ssh-import-id==5.11
starlette==0.20.4
systemd-python==234
tabulate==0.8.10
telegram-send==0.37
termcolor==2.4.0
tesserocr==2.6.3
texttable==1.6.4
tifffile==2024.2.12
tlslite-ng==0.7.6
tomli==2.0.1
tornado==6.2
twine==4.0.2
typing==3.7.4.3
tzdata==2022.5
tzlocal==4.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.5
uvicorn==0.18.3
vboxapi==1.0
virtualenv==20.17.1
voluptuous==0.13.1
wadllib==1.3.6
webencodings==0.5.1
websocket-client==1.2.3
-e git+https://github.com/WoeUSB/WoeUSB-ng.git@18e8918f75af26c0258a5b5f7bdb13acb76611eb#egg=WoeUSB_ng
wrapt==1.16.0
wsproto==1.1.0
wxPython==4.0.7
xdg==5
xkit==0.0.0
xlib==0.21
xlrd==1.2.0
xmltodict==0.12.0
youtube-dl @ file:///home/sanjeev/Downloads/youtube-dl
zipp==1.0.0
alembic==1.13.2
altair==5.3.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.2
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
streamlit==1.37.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
